Username field at registration + profile update page
Able to update username, and other details, as well as uploading profile image through firebase storage.

Only issue here is that changing the email causes the session to timeout since the bearer token is no longer "valid" for the previous email. But can relogin using the new email just fine.